requirements.txt
matplotlib==2.1.2
pandas==0.22.0
seaborn==0.8.1
progressbar2==3.34.3
